update usersrole system done

+++ lib/redux/api/rolesSupaApi.ts
@@ -1,16 +1,34 @@
 import { createApi, fakeBaseQuery } from "@reduxjs/toolkit/query/react";
 import supaClientHandler from "@/lib/Supa/SupaClient";
-import { Role } from "@/types";
+import { Role, User } from "@/types";
+
+function getUniquePermissions(roles: Omit<Role, "name">[]): string[] {
+  const permissionsSet = new Set<string>();
+
+  roles?.forEach((roleData) => {
+    const permissions = roleData.permissions;
+    if (permissions && Array.isArray(permissions)) {
+      permissions.forEach((permission) => {
+        if (permission) {
+          permissionsSet.add(permission);
+        }
+      });
+    }
+  });
+
+  return Array.from(permissionsSet);
+}
 
 export const rolesSupaApi = createApi({
   reducerPath: "rolesSupaApi",
   baseQuery: fakeBaseQuery(),
+  tagTypes: ["ROLES"],
   endpoints: (builder) => ({
     getRole: builder.query<Role, number>({
       queryFn: async (args, api, extraOptions, baseQuery) => {
         const supabase = supaClientHandler;
         const { data, error } = await supabase
-          .from("roles")
+          .from("Role")
           .select("*")
           .eq("id", args)
           .single();
@@ -21,6 +39,8 @@ export const rolesSupaApi = createApi({
 
         return { data: data };
       },
+      providesTags: (result, arg) =>
+        result ? [{ type: "ROLES", arg }, "ROLES"] : ["ROLES"],
     }),
     getRoles: builder.query<Role[], { page: number; perPage: number }>({
       queryFn: async (args, api, extraOptions, baseQuery) => {
@@ -29,7 +49,7 @@ export const rolesSupaApi = createApi({
         const start = perPage * page;
         const end = perPage * page + perPage;
         const { data, error } = await supabase
-          .from("roles")
+          .from("Role")
           .select("*")
           .range(start, end);
         if (error) {
@@ -38,11 +58,18 @@ export const rolesSupaApi = createApi({
 
         return { data: data };
       },
+      providesTags: (result, error, arg) =>
+        result
+          ? [
+              ...result.map(({ id }) => ({ type: "ROLES" as const, id })),
+              "ROLES",
+            ]
+          : ["ROLES"],
     }),
-    createRole: builder.mutation<any, Role>({
+    createRole: builder.mutation<any, Omit<Role, "id">>({
       queryFn: async (arg, api, extraOptions, baseQuery) => {
         const supabase = supaClientHandler;
-        const { error } = await supabase.from("roles").insert({ ...arg });
+        const { error } = await supabase.from("Role").insert({ ...arg });
 
         if (error) {
           return { error: error };
@@ -50,12 +77,13 @@ export const rolesSupaApi = createApi({
 
         return { data: undefined };
       },
+      invalidatesTags: ["ROLES"],
     }),
     updateRole: builder.mutation<any, { roleId: number; role: Partial<Role> }>({
       queryFn: async ({ roleId, role }, api, extraOptions, baseQuery) => {
         const supabase = supaClientHandler;
         const { error } = await supabase
-          .from("roles")
+          .from("Role")
           .update(role)
           .eq("id", roleId);
 
@@ -65,20 +93,69 @@ export const rolesSupaApi = createApi({
 
         return { data: undefined };
       },
+      invalidatesTags: (result, error, { roleId }) => [
+        { type: "ROLES", id: roleId },
+        "ROLES",
+      ],
     }),
     deleteRole: builder.mutation<any, number>({
       queryFn: async (roleId, api, extraOptions, baseQuery) => {
         const supabase = supaClientHandler;
-        const { error } = await supabase
-          .from("roles")
-          .delete()
-          .eq("id", roleId);
+        const { error } = await supabase.from("Role").delete().eq("id", roleId);
 
         if (error) {
           return { error: error };
         }
 
         return { data: undefined };
+      },
+      invalidatesTags: (result, error, roleId) => [
+        { type: "ROLES", id: roleId },
+        "ROLES",
+      ],
+    }),
+    getUserRoles: builder.query<
+      { Role: Role | null; User: Pick<User, "uid"> | null }[] | null,
+      void
+    >({
+      queryFn: async (arg, api, extraOptions, baseQuery) => {
+        const supabase = supaClientHandler;
+        const { data: userData, error: getUserError } =
+          await supabase.auth.getUser();
+        if (getUserError) {
+          return { error: getUserError };
+        }
+        const { data, error } = await supabase
+          .from("UserRole")
+          .select("Role!inner(*),User!inner(uid)")
+          .eq("users.uid", userData.user.id);
+        if (error) {
+          return { error: error };
+        }
+        const rolesData = data;
+        return { data: rolesData || [] };
+      },
+    }),
+    getUserPermissions: builder.query<string[], void>({
+      queryFn: async (arg, api, extraOptions, baseQuery) => {
+        const supabase = supaClientHandler;
+        const { data: userData, error: getUserError } =
+          await supabase.auth.getUser();
+        if (getUserError) {
+          return { error: getUserError };
+        }
+        const { data, error } = await supabase
+          .from("User")
+          .select("UserRole(Role(permissions))")
+          .eq("uid", userData.user.id)
+          .single();
+        if (error) {
+          return { error: error };
+        }
+
+        const rolesData: Omit<Role, "name">[] = data.UserRole as any;
+
+        return { data: getUniquePermissions(rolesData) || [] };
       },
     }),
   }),
@@ -90,4 +167,5 @@ export const {
   useCreateRoleMutation,
   useUpdateRoleMutation,
   useDeleteRoleMutation,
+  useGetUserPermissionsQuery,
 } = rolesSupaApi;
